Added php that flows reviews in proper order in the wall/timeline
When I clicked on a show it didn’t always appear at the top of the
wall/timeline. That was because the tv show id numbers (from the api)
are random which is useless for ordering purposes. So I am trying to
create a method that generates a reliable increment. NOTE: At the
moment the increment is triggered (counts down) each time the page
refreshes.